Add new styling to sections
