+++ SharePointFramework/ProjectWebParts/src/components/ProjectInformation/Actions/Actions.module.scss
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   gap: 5px;
+  padding-top: 16px;
 
   .button {
     justify-content: flex-start;
