+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/PublishedStatus/PublishedStatus.tsx
@@ -4,10 +4,12 @@ import React, { FC } from 'react'
 import { useProjectStatusContext } from '../context'
 import styles from '../ProjectStatus.module.scss'
 import { CheckmarkSquare24Filled } from '@fluentui/react-icons'
+import { Shimmer } from '@fluentui/react'
 
 export const PublishedStatus: FC = () => {
   const { state } = useProjectStatusContext()
   return (
+    <Shimmer isDataLoaded={state.isDataLoaded}>
     <div className={styles.publishedStatus}>
       <div className={styles.publishedStatusIcon} hidden={!state.selectedReport?.published}>
         <CheckmarkSquare24Filled />
@@ -19,6 +21,7 @@ export const PublishedStatus: FC = () => {
             : strings.NotPublishedStatusReport
         }
       />
-    </div>
+      </div>
+    </Shimmer>
   )
 }
