+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/BaseSection/BaseSection.module.scss
@@ -1,26 +1,24 @@
-@import '~@fluentui/react/dist/sass/References.scss';
+@import "~@fluentui/react/dist/sass/References.scss";
 
 .baseSection {
-  @include ms-Grid;
-  margin: 30px 0 0 0;
-  background-color: #eff6fa;
-  border-top: 5px solid #035484;
+  box-shadow: var(--shadow4);
+  border-radius: var(--borderRadiusMedium);
 
   &.transparent {
-    background-color: transparent !important;
-    border-color: transparent;
+    box-shadow: none;
+    border-radius: none;
   }
 
   &.noPadding .container {
     padding: 0 !important;
   }
 
-  &.noMargin, &.noMargin .container {
+  &.noMargin,
+  &.noMargin .container {
     margin: 0 !important;
   }
 
   .container {
-    @include ms-Grid-row;
-    padding: 15px 30px 30px 30px;
+    padding: 28px;
   }
 }
