+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/Sections.module.scss
@@ -1,6 +1,8 @@
 @import '~@fluentui/react/dist/sass/References.scss';
 
 .root {
-  padding: 0;
-  margin: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 28px;
+  padding-top: 28px;
 }
