+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/SummarySection/SummarySection.tsx
@@ -42,25 +42,29 @@ export const SummarySection: FC<ISummarySectionProps> = (props) => {
 
   return (
     <BaseSection {...props}>
-      <div className={styles.root}>
+      <div className={styles.summarySection}>
         {props.showProjectInformation && (
           <div className={styles.projectInformation}>
             <ProjectInformation
               {...pick(context.props, 'siteId', 'webAbsoluteUrl', 'spfxContext')}
               page='ProjectStatus'
-              hideAllActions={true}
-              hideStatusReport={true}
+              hideAllActions
+              hideStatusReport
             />
           </div>
         )}
-        <div
-          className={className([
-            styles.statusElements,
-            !props.showProjectInformation && styles.fullWidth
-          ])}
-        >
+        <div className={className([
+          styles.statusElements,
+          !props.showProjectInformation && styles.fullWidth
+        ])}>
           <div className={styles.container} dir='ltr'>
-            <div className={styles.row}>{renderStatusElements()}</div>
+            <div className={className([
+              styles.row,
+              props.iconsOnly ? styles.iconsOnly : ''
+            ])}
+            >
+              {renderStatusElements()}
+            </div>
           </div>
         </div>
       </div>
