--- SharePointFramework/ProjectWebParts/src/components/ProjectStatus/StatusElement/StatusElementIcon/StatusElementIcon.module.scss
@@ -1,7 +1,0 @@
-@import '~@fluentui/react/dist/sass/References.scss';
-
-.root {
-  @include ms-Grid-col;
-  @include ms-sm12;
-  @include ms-md2;
-}
