+++ SharePointFramework/ProjectWebParts/src/components/ProjectTimeline/TimelineList/TimelineList.module.scss
@@ -3,12 +3,6 @@
 .timelineList {
   padding: 32px;
 
-  .stickyHeader {
-    .headerColumns {
-      position: inherit;
-    }
-  }
-
   .commandBar {
     height: 42px;
     margin: 0px auto;
