+++ SharePointFramework/ProjectWebParts/src/loc/nb-no.js
@@ -80,14 +80,14 @@ define([], function () {
     DocumentsListName: 'Dokumenter',
     EditItemLabel: 'Rediger',
     EditProjectInformationText: 'Rediger prosjektinformasjon',
-    EditReportButtonLabel: 'Rediger status',
+    EditReportButtonLabel: 'Rediger',
     EditReportButtonDescription: 'Rediger den valgte statusrapporten.',
     EditReportButtonDescriptionNoPermission: 'Du har ikke tilgang til å redigere denne statusrapporten.',
     EditSiteInformationText: 'Rediger områdeinformasjon',
     ExternalUsersConfigInfoText: 'Her konfigueres felter/kolonner som skal vises for eksterne brukere/gjester.<br /><br />Felter/kolonner som skal vises for brukere i organisasjonen konfigureres på porteføljeområdet av administrator.',
     ExternalUsersConfigText: 'Konfigurasjon',
     GeneralGroupName: 'Generelt',
-    GetSnapshotButtonLabel: 'Åpne som øyeblikksbilde',
+    GetSnapshotButtonLabel: 'Åpne øyeblikksbilde',
     GetSnapshotButtonDescription: 'Åpner den valgte statusrapporten som et øyeblikksbilde.',
     GtModerationStatus_Choice_Draft: 'Kladd',
     GtModerationStatus_Choice_Published: 'Publisert',
@@ -131,7 +131,7 @@ define([], function () {
     MoveOn: 'Gå videre',
     NavigateToSourceUrlText: 'Gå tilbake',
     NewItemLabel: 'Nytt element',
-    NewStatusReportLabel: 'Ny statusrapport',
+    NewStatusReportLabel: 'Opprett',
     NewStatusReportDescription: 'Opprett en ny statusrapport for prosjektet.',
     NewStatusReportDescriptionNoPermission: 'Du har ikke tilgang til å opprette en ny statusrapport.',
     NewStatusReportTitle: 'Ny statusrapport for {0}',
