+++ SharePointFramework/shared-library/src/components/Toolbar/Toolbar.module.scss
@@ -1,6 +1,7 @@
 .root {
   display: flex;
   justify-content: space-between;
+  gap: 15px;
 
   .toolbar {
     gap: 15px;
